refactoring: 신고하기 MVVM으로 변경
